ENH: Add version switcher json to the base documentation site
This makes the release process a bit more clear, since the
``versions.json`` file should naturally be kept in sync with the
``index.html`` file in the docs base repository/site.

The full "process" of moving is adding the file here, and then
making a PR against NumPy main to fix the links and remove the
(now unnecessary) copy from there.